Created HomeFragmentTest with initial test launching the fragment and checking the appBar and its title.
- Setup to launch fragment when using Hilt:
Explanation: When using Hilt, it is necessary to manually define the activity from which the fragments will be launched (activity that is required to be annotated with AndroidEntryPoint in Hilt). When using launchFragmentInContainer it uses a default empty activity, so it is necessary to create an util function to use a custom HiltTestActivity annotated with @AndroidEntryPoint.
Steps:
- Add launchFragmentInHiltContainer util function from https://github.com/android/architecture-samples/blob/dev-hilt/app/src/androidTest/java/com/example/android/architecture/blueprints/todoapp/HiltExt.kt#L38
- Create HiltTestActivity and associated manifest in a debug variant.

- Apart from the required setUp to launched fragment when using Hilt, another best practice is to use Robots, which are in turn based on builder pattern.
Because of this I created ScreenRobot and HomeRobot and use them to launch fragment in HomeFragment and to check the appBar and its title.