Refactoring: TestModules, BaseHiltTest and TestMoviesApi fix.
Moved out Local and Network Modules from the class, so that they can be used with TestModule and in all classes which is recommended as best practice.

Created BaseHiltTest to inject setup and database setup (should be closed on every test completion).

Fix: Changed FakeMoviesApi from object to class, so that different Apis will be used on each test.